pyproject.toml
[project]
name = "gemini-sre-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
  "pyyaml>=6.0.1",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.0.0",
  "click>=8.0.0",

  "tenacity>=8.2.3",
  "slowapi>=0.1.9",
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "httpx>=0.27.0",
  "google-cloud-logging>=3.10.0",
  "google-cloud-pubsub>=2.21.1",
  "google-cloud-aiplatform>=1.53.0",
  "aiofiles>=23.2.1",
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.7",
  "PyGithub>=2.3.0",
  "google-auth>=2.29.0",
  "google-auth-oauthlib>=1.2.0",
  "google-auth-httplib2>=0.2.0",
  "google-generativeai>=0.3.0",
  "circuitbreaker>=2.1.3",
  "limits>=3.5.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
markers = [
  "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
